+++ tests/integration/cases/call_deep_dotted_method/ObjectiveC_call.m
@@ -1,13 +1,11 @@
 #import <Foundation/Foundation.h>
-#pragma clang diagnostic ignored "-Wstrict-prototypes"
-#pragma clang diagnostic ignored "-Wdeprecated-non-prototype"
-static void obj_api_client_post_stub_() {}
-struct clientType_ { void (*post)(); };
+static void obj_api_client_post_stub_(id _a0) { (void)_a0; }
+struct clientType_ { void (*post)(id); };
 struct apiType_ { struct clientType_ client; };
 struct objType_ { struct apiType_ api; };
 static const struct objType_ obj = { .api = { .client = { .post = obj_api_client_post_stub_ } } };
 void check_(void) {
 obj.api.client.post(@"hello");
-obj.api.client.post(42);
+obj.api.client.post(@(42));
 obj.api.client.post(@YES);
 }