+++ tests/integration/cases/dict_mixed_int_widths/CSharp.cs
@@ -1,0 +1,6 @@
+using System.Collections.Generic;
+var my_data = new Dictionary<string, object> {
+    ["a"] = 1,
+    ["b"] = 3000000000,
+    ["c"] = "x"
+};