+++ tests/integration/cases/dict_mixed_int_widths/Erlang.erl
@@ -1,0 +1,9 @@
+-module(check).
+-export([x/0]).
+x() ->
+    My_data = #{
+        "a" => 1,
+        "b" => 3000000000,
+        "c" => "x"
+    },
+    My_data.