+++ tests/integration/cases/dict_mixed_int_widths/Racket.rkt
@@ -1,0 +1,6 @@
+#lang racket
+(hash
+    "a" 1
+    "b" 3000000000
+    "c" "x"
+)