+++ tests/integration/cases/dict_mixed_int_widths/Scala_combined.scala
@@ -1,0 +1,12 @@
+object Check {
+var my_data = Map(
+    "a" -> 1,
+    "b" -> 3000000000L,
+    "c" -> "x",
+)
+my_data = Map(
+    "a" -> 1,
+    "b" -> 3000000000L,
+    "c" -> "x",
+)
+}